test: Fix React 18.3 findDOMNode warning (#48958)
* fix: try to fix React 18.3 warning

* chore: back of ci

* chore: fix type

* chore: fix lint

* chore: update deps

* fix: component error

* test: update snapshot

* test: update snapshot

* test: update snapshot

* test: update snapshot

* test: update snapshot

* chore: update deps

* test: update snapshot

* test: update snapshot

* test: update snapshot

* chore: update deps

* test: update snapshot

* test: update snapshot

* test: update snapshot

* chore: bump version
