+++ components/checkbox/index.zh-CN.md
@@ -68,10 +68,11 @@ interface Option {
 
 #### Checkbox
 
-| 名称    | 描述     | 版本 |
-| ------- | -------- | ---- |
-| blur()  | 移除焦点 |      |
-| focus() | 获取焦点 |      |
+| 名称          | 描述                      | 版本   |
+| ------------- | ------------------------- | ------ |
+| blur()        | 移除焦点                  |        |
+| focus()       | 获取焦点                  |        |
+| nativeElement | 返回 Checkbox 的 DOM 节点 | 5.17.3 |
 
 ## 主题变量（Design Token）
