+++ components/tour/__tests__/demo-extend.test.ts
@@ -1,3 +1,14 @@
 import { extendTest } from '../../../tests/shared/demoTest';
 
-extendTest('tour');
+jest.mock('../.', () => {
+  const OriReact = jest.requireActual('react');
+  const OriTour = jest.requireActual('../.').default;
+
+  const ProxyTour = OriReact.forwardRef((props: any, ref: any) =>
+    OriReact.createElement(OriTour, { ...props, open: true, ref }),
+  );
+
+  return ProxyTour;
+});
+
+extendTest('tour', { skip: ['render-panel.tsx'] });
