+++ package.json
@@ -63,7 +63,7 @@
     "format": "biome format --write .",
     "install-react-16": "npm i --no-save --legacy-peer-deps react@16 react-dom@16 @testing-library/react@12",
     "install-react-17": "npm i --no-save --legacy-peer-deps react@17 react-dom@17 @testing-library/react@12",
-    "install-react-18": "npm i --no-save --legacy-peer-deps react@18.2 react-dom@18.2",
+    "install-react-18": "npm i --no-save --legacy-peer-deps react@18 react-dom@18",
     "prelint": "dumi setup",
     "lint": "npm run version && npm run tsc && npm run lint:script && npm run lint:demo && npm run lint:md && npm run lint:style && npm run lint:changelog",
     "lint-fix": "npm run lint-fix:script && npm run lint-fix:demo",
@@ -124,24 +124,24 @@
     "@ctrl/tinycolor": "^3.6.1",
     "@rc-component/color-picker": "~1.5.3",
     "@rc-component/mutate-observer": "^1.1.0",
-    "@rc-component/tour": "~1.14.2",
+    "@rc-component/tour": "~1.15.0",
     "@rc-component/trigger": "^2.1.1",
     "classnames": "^2.5.1",
     "copy-to-clipboard": "^3.3.3",
     "dayjs": "^1.11.10",
     "qrcode.react": "^3.1.0",
-    "rc-cascader": "~3.25.0",
-    "rc-checkbox": "~3.2.0",
+    "rc-cascader": "~3.26.0",
+    "rc-checkbox": "~3.3.0",
     "rc-collapse": "~3.7.3",
     "rc-dialog": "~9.4.0",
     "rc-drawer": "~7.1.0",
     "rc-dropdown": "~4.2.0",
     "rc-field-form": "~2.0.1",
     "rc-image": "~7.6.0",
-    "rc-input": "~1.4.5",
-    "rc-input-number": "~9.0.0",
-    "rc-mentions": "~2.11.1",
-    "rc-menu": "~9.13.0",
+    "rc-input": "~1.5.0",
+    "rc-input-number": "~9.1.0",
+    "rc-mentions": "~2.13.1",
+    "rc-menu": "~9.14.0",
     "rc-motion": "^2.9.1",
     "rc-notification": "~5.4.0",
     "rc-pagination": "~4.0.4",
@@ -150,16 +150,16 @@
     "rc-rate": "~2.12.0",
     "rc-resize-observer": "^1.4.0",
     "rc-segmented": "~2.3.0",
-    "rc-select": "~14.13.4",
+    "rc-select": "~14.14.0",
     "rc-slider": "~10.6.2",
     "rc-steps": "~6.0.1",
     "rc-switch": "~4.1.0",
     "rc-table": "~7.45.6",
-    "rc-tabs": "~15.0.0 ",
-    "rc-textarea": "~1.6.3",
+    "rc-tabs": "~15.1.0",
+    "rc-textarea": "~1.7.0",
     "rc-tooltip": "~6.2.0",
     "rc-tree": "~5.8.7",
-    "rc-tree-select": "~5.20.0",
+    "rc-tree-select": "~5.21.0",
     "rc-upload": "~4.5.2",
     "rc-util": "^5.40.1",
     "scroll-into-view-if-needed": "^3.1.0",
